requirements.txt
Flask==2.3.3
flask-restx==1.3.1
PyMuPDF==1.23.8
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.4
torch==2.0.1
torchvision==0.15.2
transformers==4.35.2
ultralytics==8.0.206
Werkzeug==2.3.7
python-dotenv==1.0.0